Move autocorrection function into its own file
