+++ pytest.ini
@@ -20,6 +20,7 @@ markers =
     graceful_error: 'graceful_error' tests
     database_manager: database tests
     galaxy: marks a test as a galaxy test
+    autocorrect: tests for the autocorrect util
 testpaths = tests
 
 addopts = --doctest-modules
